+++ docs/doxygen/html/search/all_4.js
@@ -5,6 +5,7 @@ var searchData=
   ['displaycriteriadate',['displayCriteriaDate',['../classphp__bs__grid.html#abda0b147b4a556bb1f78711f90617cce',1,'php_bs_grid']]],
   ['displaycriterialookup',['displayCriteriaLookup',['../classphp__bs__grid.html#a48fcff4b71851699f42fcd0566df4bb5',1,'php_bs_grid']]],
   ['displaycriteriamultiselectcheckbox',['displayCriteriaMultiselectCheckbox',['../classphp__bs__grid.html#a98634e8fb2c26437bc90e3e488da7e58',1,'php_bs_grid']]],
+  ['displaycriterianumber',['displayCriteriaNumber',['../classphp__bs__grid.html#a22d860e32a59ccc5bf63f6394fd6bf11',1,'php_bs_grid']]],
   ['displaycriteriatext',['displayCriteriaText',['../classphp__bs__grid.html#a88e0e7a16b76d3299b04d198080ff388',1,'php_bs_grid']]],
   ['displayrowsperpage',['displayRowsPerPage',['../classphp__bs__grid.html#a04b0e8f58dfb58282c3ec703f8c8c7bd',1,'php_bs_grid']]],
   ['displaytabledata',['displayTableData',['../classphp__bs__grid.html#a2ad3c171e07b2eb169751c7373593ff2',1,'php_bs_grid']]],