+++ CHANGELOG.md
@@ -10,9 +10,10 @@ Project page https://github.com/pontikis/php_bs_grid
 License MIT https://raw.github.com/pontikis/php_bs_grid/master/MIT_LICENSE
 
 
-Release 0.9.5 (XX July 2017)
+Release 0.9.5 (02 July 2017)
 --------------------------
 
+* add operators to multiselect_checkbox (one or more of, not given) #22
 * add operators to autocomplete (equal, not given) #23
 * add not equal operator to lookup filters #27
 * ability to append/prepend element in filters #26